+++ src/components/pages/Home.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Hero from '../sections/Hero';
+import Hero from '../sections/Hero/Hero';
 
 function Home() {
   return (
